refactor     move configs to `config` package
